refactor: move files to domain
- LottoGenerator
- LottoRanker
- Validator
